Move out tests for time_format.
